requirement.txt
aiodns==3.0.0
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
aiosqlite==0.17.0
amqp==5.1.1
appdirs==1.4.4
apt-xapian-index==0.49
apturl==0.5.2
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
attrs==22.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
behave==1.2.6
beniget==0.4.1
billiard==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.0
cairocffi==1.2.0
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
command-not-found==0.3
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debtags==2.1
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.2
djangorestframework==3.14.0
dot2tex==2.11.3
duplicity==0.8.21
exceptiongroup==1.1.1
execnet==1.9.0
fasteners==0.18
filelock==3.11.0
fonttools==4.29.1
frozenlist==1.2.0
fs==2.4.12
future==0.18.2
gast==0.5.2
GDAL==3.4.1
Genshi==0.7.6
gmpy2==2.1.2
google-api-core==2.11.0
google-api-python-client==1.7.9
google-auth==2.17.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
googleapis-common-protos==1.59.0
gyp==0.1
h11==0.14.0
hiredis==1.0.1
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.3.0
incremental==21.3.0
iniconfig==2.0.0
ipython==7.31.1
jaraco.classes==3.2.3
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
kazam==1.4.5
keyring==23.13.1
kiwisolver==1.3.2
kombu==5.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==0.0.0
multidict==5.1.0
mysqlclient==2.2.0
netifaces==0.11.0
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
pandas==2.0.0
parameterized==0.9.0
paramiko==2.9.3
parse==1.19.0
parse-type==0.6.0
parso==0.8.1
pdbp==1.2.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
postfix-mta-sts-resolver==1.1.2
prompt-toolkit==3.0.39
protobuf==4.22.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.20.1
pycares==4.1.2
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
PyDrive==1.3.1
PyGetWindow==0.0.9
Pygments==2.15.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynose==1.4.2
PyOpenGL==3.1.5
pyOpenSSL==23.1.1
pyotp==2.8.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
pyserial==3.5
PySocks==1.7.1
pytest==7.3.0
pytest-forked==1.6.0
pytest-html==2.0.1
pytest-metadata==2.0.4
pytest-ordering==0.6
pytest-rerunfailures==11.1.2
pytest-xdist==3.2.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-pam==1.8.4
python3-xlib==0.15
pythran==0.10.0
pytweening==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
pyzor==1.0.0
redis==4.6.0
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.3
rpy2==3.4.5
rsa==4.9
sbvirtualdisplay==1.2.0
scipy==1.8.0
scour==0.38.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.8.3
seleniumbase==4.13.24
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.2
sympy==1.9
systemd-python==234
tabcompleter==1.1.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.1.0
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==3.0.1
urllib3==1.26.15
usb-creator==0.3.7
uvloop==0.16.0
vine==5.0.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==3.8.6
webencodings==0.5.1
wsproto==1.2.0
wxPython==4.0.7
xcffib==0.11.1
xdg==5
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.1.0
yarl==1.7.2
zipp==3.15.0
zope.interface==5.4.0
